requirements.txt
alembic==1.15.2
click==8.1.8
dominate==2.9.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
